==> 04.de_workflow_day8.Rmd <==
---
title: "DE analysis - Day8"
author: "Sergey Naumenko"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: hide
    df_print: paged
    highlights: pygments
    number_sections: yes
    self_contained: yes
    theme: default
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: yes
  pdf_document:
    toc: yes
---

# Overview
- Principal Investigator: Beth Overmoyer
- Experiment: RNAseq_analysis_of_inflammatory_breast_cancer_hbc04141
- study 6 was excluded because if low read depth in 3373-3
- https://www.bioconductor.org/packages/release/bioc/vignettes/DEGreport/inst/doc/DEGreport.html
- [AnnotationHub](https://bioconductor.org/packages/release/bioc/html/AnnotationHub.html). We use ensembl version matching bcbio pipeline - v94.
- [HBC materials](https://hbctraining.github.io/DGE_workshop_salmon/lessons/03_DGE_QC_analysis.html)
- [HBC materials - functional analysis](https://hbctraining.github.io/DGE_workshop_salmon/lessons/functional_analysis_2019.html)
- http://bioconductor.org/packages/release/bioc/vignettes/DESeq2/inst/doc/DESeq2.html
- this is DE for Day8 samples

```{r include = FALSE}
remove_cases_2_19 <- FALSE
rebuild_rds <- FALSE
## Setup
### Bioconductor and CRAN libraries used

library(DESeq2)
library(tidyverse)
library(RColorBrewer)
library(pheatmap)
library(DEGreport)
library(tximport)
library(ggplot2)
library(ggrepel)
library(knitr)
library(AnnotationHub)
library(ensembldb)
library(org.Hs.eg.db)
library(clusterProfiler)
library(writexl)

ggplot2::theme_set(theme_light(base_size = 14))

opts_chunk[["set"]](
    cache = FALSE,
    dev = c("png", "pdf"),
    error = TRUE,
    highlight = TRUE,
    message = FALSE,
    prompt = FALSE,
    tidy = FALSE,
    warning = FALSE)
```
```{r include = FALSE}
# Have a folder called `data`, and copy your Salmon folders here from the cluster. 
## List all directories containing data  
### change the pattern to something specific to your Salmon folders
samples <- list.files(path = "./data/final", 
                      full.names = T,
                      pattern = "^S")

## Obtain a vector of all filenames including the path
files <- file.path(samples, "salmon", "quant.sf")
files

## Since all quant files have the same name it is useful to have names for each element
### change the string in str_replace so the pattern matches your filenames
names(files) <- str_replace(samples, "./data/final/", "") 
```

```{r include = FALSE}
# Load the data and metadata
meta <- read_csv("tables/metadata_corrected.csv") %>% 
  column_to_rownames(var = "samplename") %>% 
  dplyr::filter(treatment == "post") %>% 
  drop_na(response)

if (remove_cases_2_19){
   meta <- meta %>% dplyr::filter(!study_id %in% c(2, 19))
   files <- files[rownames(meta)]
}

meta$response <- str_replace(meta$response, "Yes", "pCR")
meta$response <- str_replace(meta$response, "No", "non-pCR")

protein_coding_genes <- read_csv("tables/ensembl_w_description.protein_coding.csv")
```
```{r include = FALSE}
annotation_file <- "data/ensembl94_hg38_annotations.txt"
gene_symbol_file <- "data/gene_symbol.txt"
if (file.exists(annotation_file)){
    hsdb <- read_tsv(annotation_file)
    gene_symbol <- read_tsv(gene_symbol_file)
}else{
# Connect to AnnotationHub
ah <- AnnotationHub()

# Query AnnotationHub
hs_ens <- query(ah, c("Homo sapiens", "EnsDb"))

# Get Ensembl94 - used in bcbio
hs_ens <- hs_ens[["AH64923"]]

# Extract gene-level information
txdb <- transcripts(hs_ens,
                    return.type = "data.frame") %>% 
  dplyr::select(tx_id, gene_id)  

genedb <- genes(hs_ens, 
                return.type = "data.frame") %>% 
  dplyr::select(gene_id, gene_name, symbol)  

gene_symbol <- genedb %>% dplyr::select(gene_id, symbol)

write_tsv(gene_symbol, gene_symbol_file)
hsdb <- inner_join(txdb, genedb)
write.table(hsdb, 
            file = "data/ensembl94_hg38_annotations.txt", 
            sep = "\t", 
            row.names = F, 
            quote = F)
}
```

```{r include = TRUE}
# Read in  a tx2gene file with transcript identifiers in the first column and gene identifiers in the second column
tx2gene <- hsdb[, c("tx_id", "gene_id")]


# Run tximport
txi_file <- "data/txi.day8.RDS"
if (!rebuild_rds & file.exists(txi_file)){
    txi <- readRDS(txi_file)
}else{
    files <- files[rownames(meta)]
    txi <- tximport(files, 
                type = "salmon", 
                tx2gene = tx2gene, 
                countsFromAbundance = "lengthScaledTPM", 
                ignoreTxVersion = FALSE)
    saveRDS(txi, txi_file)
}

# Look at the counts
class(txi)
attributes(txi)
txi$counts %>% View()
```

# Checking to see that the transcript to gene mapping is correct
When you have annotations that are from a different source from your reference you can run into problems (i.e lose genes). Some checks you can do before proceeding:

1. Look at the dimensions of your count matrix. Do you have ~20k genes present? `dim(txi$counts)`
2. When running `tximport()` you will get a message in your console. If you see something like `transcripts missing from tx2gene` start troubleshooting.
```{r}
dim(txi$counts)
```

# Sanity check that metadata matches your expression
It is always a good idea to check if:

1. Do you have expression data for all samples listed in your metadata?
2. Are the samples in your expression data in the same order as your metadata?
```{r}
### Check that sample names match in both files
all(colnames(txi$counts) %in% rownames(meta))

### Check that sample names match in both files
all(colnames(txi$counts) %in% rownames(meta))

### Check that all samples are in the same order
meta <- meta[colnames(txi$counts),]
all(colnames(txi$counts) == rownames(meta))

```

# Run DESeq2
```
estimating size factors
estimating dispersions
gene-wise dispersion estimates
mean-dispersion relationship
final dispersion estimates
fitting model and testing
```

* [Estimating size factors and count normalization](https://hbctraining.github.io/DGE_workshop_salmon/lessons/02_DGE_count_normalization.html)
* [Gene-wise dispersions](https://hbctraining.github.io/DGE_workshop_salmon/lessons/04_DGE_DESeq2_analysis.html)
* [Mean-dispersion(variance) relationship and the Negative Binomial Model](https://hbctraining.github.io/DGE_workshop_salmon/lessons/01_DGE_setup_and_overview.html#modeling-count-data)
* [Model fitting and hypothesis testing](https://hbctraining.github.io/DGE_workshop_salmon/lessons/05_DGE_DESeq2_analysis2.html)


# Wald test
*Here we subset protein coding genes*.
```{r include = TRUE}
## Create DESeq2Dataset object
dds_file <- "data/dds.day8.RDS"
meta$treatment <- as.factor(meta$treatment)
meta$response <- as.factor(meta$response)
meta$er <- as.factor(meta$er)
meta$date_of <- as.factor(meta$date_of)
meta$tumor_percentage <- as.factor(meta$tumor_percentage)
meta$tumor_percentage_high <- as.factor(meta$tumor_percentage_high)
meta$tumor_percentage_high <- relevel(meta$tumor_percentage_high, "low")


if (remove_cases_2_19){
   non_responders <- meta %>% dplyr::filter(study_id %in% c(2, 19)) %>% row.names() 
}

if (!rebuild_rds & file.exists(dds_file)){
    dds <- readRDS(dds_file)
}else{
    dds <- DESeqDataSetFromTximport(txi,
                                colData = meta, 
                                design = ~response)
    
    if (remove_cases_2_19){
        dds <- dds[,!colnames(dds) %in% non_responders]
    }
    
    design(dds) <- formula(~response + er + tumor_percentage_high + date_of)
    
    # subset protein-coding genes
    pc_genes <- intersect(protein_coding_genes$ensembl_gene_id, row.names(dds))
    dds <- dds[pc_genes,]
    # 100 reads / 20 samples
    keep <- rowSums(counts(dds)) >= 100
    dds <- dds[keep,]

    # Run DESeq2
    dds <- DESeq(dds)
    saveRDS(dds, dds_file)
}
```

# DEGreport QC

## Size factor QC - samples 1-20
```{r fig.width=10, fig.height = 10}
counts <- counts(dds, normalized = TRUE)
design <- as.data.frame(colData(dds))
degCheckFactors(counts[, 1:20])

```

# Mean-Variance QC plots

## response
```{r fig.width=15, fig.height = 15}
res <- results(dds)
degQC(counts, design[["response"]], pvalue = res[["pvalue"]])
```

## ER
```{r fig.width=15, fig.height = 15}
degQC(counts, design[["er"]], pvalue = res[["pvalue"]])
```

## tumor_percentage_high
```{r fig.width=15, fig.height = 15}
degQC(counts, design[["tumor_percentage_high"]], pvalue = res[["pvalue"]])
```

# Covariates effect on count data
```{r}
mdata <- colData(dds) %>%  as.data.frame()  %>% 
  dplyr::select(response, er, date_of, tumor_percentage_high)

#resCov <- degCovariates(log2(counts(dds)+0.5), mdata)
```

```{r}
mdata %>% ggplot(aes(tumor_percentage_high, fill = response)) + geom_bar(position = "dodge2")
```

# Covariates correlation with metrics
```{r}
cor <- degCorCov(mdata)
```


```{r}
mdata %>% ggplot(aes(date_of, fill = response)) + geom_bar(position = "dodge2")
```

# Sample-level QC analysis 

```{r include = TRUE}
### Transform counts for data visualization (unsupervised analysis)
rld_file <- "data/rld.day8.RDS"
if (!rebuild_rds & file.exists(rld_file)){
    rld <- readRDS(rld_file)
}else{
    rld <- rlog(dds, blind = TRUE)
    saveRDS(rld, rld_file)
}
class(rld) # what type of object is this

# we also need just a matrix of transformed counts
rld_mat <- assay(rld)
```

\newpage

## PCA - response
```{r, fig.width = 10, fig.height = 10}
# Use the DESeq2 function
plotPCA(rld, intgroup = c("response")) + geom_label_repel(aes(label = name))
```

\newpage

## PCA - ER
```{r, fig.width = 10, fig.height = 10}
# Use the DESeq2 function
plotPCA(rld, intgroup = c("er")) + geom_label_repel(aes(label = name))
```

\newpage

## PCA - tumor_percentage
```{r, fig.width = 10, fig.height = 10}
# Use the DESeq2 function
plotPCA(rld, intgroup = c("tumor_percentage")) + geom_label_repel(aes(label = name))
```

\newpage

## PCA - tumor_percentage_high
```{r, fig.width = 10, fig.height = 10}
# Use the DESeq2 function
plotPCA(rld, intgroup = c("tumor_percentage_high")) + geom_label_repel(aes(label = name))
```

\newpage

## PCA - date_of
```{r, fig.width = 10, fig.height = 10}
# Use the DESeq2 function
plotPCA(rld, intgroup = c("date_of")) + geom_label_repel(aes(label = name))
```

\newpage

# Inter-correlation analysis

## Without study_id

```{r, fig.width = 20, fig.height = 20}
# Correlation matrix
rld_cor <- cor(rld_mat)

meta$study_id <- as.factor(meta$study_id)
# Create annotation file for samples
annotation <- meta[, c("response", "er", "tumor_percentage_high", "date_of")]

# Change colors
heat.colors <- brewer.pal(6, "Blues")

# Plot heatmap
pheatmap(rld_cor, 
         annotation = annotation, 
         border = NA,
         fontsize = 20)
```

\newpage

## With study_id

```{r, fig.width = 20, fig.height = 20}
# Correlation matrix
rld_cor <- cor(rld_mat)

meta$study_id <- as.factor(meta$study_id)
# Create annotation file for samples
annotation <- meta[, c("response", "er", "tumor_percentage_high", "date_of", "study_id")]

# Change colors
heat.colors <- brewer.pal(6, "Blues")

# Plot heatmap
pheatmap(rld_cor, 
         annotation = annotation, 
         border = NA,
         fontsize = 20)
```


# Response pCR vs non-pCR for Day 8 - see Table13
```{r, include = FALSE}
# return mean counts for a group of sample in a column
get_counts_for_samples <- function(samples, column_name){
    tpm_counts <- txi$abundance %>%
        as.data.frame() %>% 
        dplyr::select(any_of(samples)) %>% 
        rowMeans() %>% 
        as.data.frame() %>% 
        rownames_to_column(var = "ensembl_gene_id")
    colnames(tpm_counts) <- c("ensembl_gene_id", "tpm")
    
    tpm_counts <- tpm_counts %>% 
        dplyr::mutate("{column_name}" := round(tpm, 2)) %>% 
        dplyr::select(-tpm)
    
    return(tpm_counts)
}

# get rid of 
comb_de_result_table <- function(results){
    results <- results %>%
        mutate(baseMean = round(baseMean, 2),
               log2FoldChange = round(log2FoldChange, 2),
               lfcSE = round(lfcSE, 2),
               stat = round(stat, 2),
               pvalue = format(pvalue, scientific = TRUE, digits = 2),
               padj = format(padj, scientific = TRUE, digits = 2))
    return(results)
}


contrast <- c("response", "pCR", "non-pCR")
resResponse <- results(dds, contrast = contrast, alpha = 0.05)
length(which(resResponse$padj < 0.05))

# Add annotations
resResponse_tb <- resResponse %>%
  data.frame() %>% 
  rownames_to_column(var = "gene") %>% 
  as_tibble() %>% 
  left_join(gene_symbol, by = c("gene" = "gene_id")) 

resResponse_tb_significant <- dplyr::filter(resResponse_tb, padj < 0.05)

samples_no <- meta %>% dplyr::filter(response == "non-pCR") %>% row.names() 
tpm_no <- get_counts_for_samples(samples_no, "non-pCR_expression_mean_tpm")

samples_yes <- meta %>% dplyr::filter(response == "pCR") %>% row.names() 
tpm_yes <- get_counts_for_samples(samples_yes, "pCR_expression_mean_tpm")

tpm_counts <- tpm_yes %>% 
    left_join(tpm_no, 
              by = c("ensembl_gene_id" = "ensembl_gene_id"))

resResponse_tb_significant <- resResponse_tb_significant %>% 
    left_join(tpm_counts, by = c("gene" = "ensembl_gene_id")) %>% 
    comb_de_result_table()

write_xlsx(list(T13.DE_response_day8 = resResponse_tb_significant), 
          "tables/T13.DE_response_day8.xlsx")


# Separate into up and down-regulated gene sets
sigResponse_up <- rownames(resResponse)[which(resResponse$padj < 0.01 & resResponse$log2FoldChange > 0)]
sigResponse_down <- rownames(resResponse)[which(resResponse$padj < 0.01 & resResponse$log2FoldChange < 0)]
```

# ER : Positive vs Negative for Day8 - Table 14
```{r, include = FALSE}
contrast <- c("er", "Positive", "Negative")
resER <- results(dds, contrast = contrast, alpha = 0.05)
length(which(resER$padj < 0.05))

# Add annotations
resER_tb <- resER %>%
  data.frame() %>% 
  rownames_to_column(var = "gene") %>% 
  as_tibble() %>% 
  left_join(gene_symbol, by = c("gene" = "gene_id")) 

resER_tb_significant <- dplyr::filter(resER_tb, padj < 0.05)

samples_pos <- meta %>% dplyr::filter(er == "Positive") %>% row.names() 
tpm_pos <- get_counts_for_samples(samples_pos, "Positive_expression_mean_tpm")

samples_neg <- meta %>% dplyr::filter(er == "Negative") %>% row.names() 
tpm_neg <- get_counts_for_samples(samples_neg, "Negative_expression_mean_tpm")

tpm_counts <- tpm_pos %>%
            left_join(tpm_neg, 
                      by = c("ensembl_gene_id" = "ensembl_gene_id"))

resER_tb_significant <- resER_tb_significant %>%
          left_join(tpm_counts, by = c("gene" = "ensembl_gene_id")) %>% 
    comb_de_result_table()


write_xlsx(list(T14.DE_ER.day8 = resER_tb_significant),
          "tables/T14.DE_ER.day8.xlsx")

# Separate into up and down-regulated gene sets
sigER_up <- rownames(resER)[which(resER$padj < 0.01 & resER$log2FoldChange > 0)]
sigER_down <- rownames(resER)[which(resER$padj < 0.01 & resER$log2FoldChange < 0)]
```

# tumor_percentage_high : High vs Low for Day8- Table 15
```{r, include = FALSE}
contrast <- c("tumor_percentage_high", "high", "low")
resTP <- results(dds, contrast = contrast, alpha = 0.05)
length(which(resTP$padj < 0.05))

# Add annotations
resTP_tb <- resTP %>%
  data.frame() %>% 
  rownames_to_column(var = "gene") %>% 
  as_tibble() %>% 
  left_join(gene_symbol, by = c("gene" = "gene_id")) 

resTP_tb_significant <- dplyr::filter(resTP_tb, padj < 0.05)

samples_high <- meta %>% dplyr::filter(tumor_percentage_high == "high") %>% row.names() 
tpm_high <- get_counts_for_samples(samples_high, "High_mean_tpm")

samples_low <- meta %>% dplyr::filter(tumor_percentage_high == "low") %>% row.names() 
tpm_low <- get_counts_for_samples(samples_low, "Low_mean_tpm")

counts_tpm <- tpm_high %>%
            left_join(tpm_low, 
                      by = c("ensembl_gene_id" = "ensembl_gene_id"))

resTP_tb_significant <- resTP_tb_significant %>%
    left_join(counts_tpm, by = c("gene" = "ensembl_gene_id")) %>% 
    comb_de_result_table()


write_xlsx(list(T15.DE_tumor_pct_high.day8 = resTP_tb_significant),
              "tables/T15.DE_tumor_percentage_high.day8.xlsx")

# Separate into up and down-regulated gene sets
sigTP_up <- rownames(resTP)[which(resTP$padj < 0.01 & resTP$log2FoldChange > 0)]
sigTP_down <- rownames(resTP)[which(resTP$padj < 0.01 & resTP$log2FoldChange < 0)]
```


# date_of: 20180323 vs 20180228 - for Day8: Table 16
```{r, include = FALSE}
contrast <- c("date_of", "20180323", "20180228")
resDO <- results(dds, contrast = contrast, alpha = 0.05)
length(which(resDO$padj < 0.05))

# Add annotations
resDO_tb <- resDO %>%
  data.frame() %>% 
  rownames_to_column(var = "gene") %>% 
  as_tibble() %>% 
  left_join(gene_symbol, by = c("gene" = "gene_id")) 

resDO_tb_significant <- dplyr::filter(resDO_tb, padj < 0.05)

samples_23 <- meta %>% dplyr::filter(date_of == "20180323") %>% row.names() 
counts_23 <- get_counts_for_samples(samples_23, "20180323_expression_mean_tpm")

samples_28 <- meta %>% dplyr::filter(date_of == "20180228") %>% row.names() 
counts_28 <- get_counts_for_samples(samples_28, "20180228_expression_mean_tpm")

tpm_counts <- counts_28 %>%
            left_join(counts_23, 
                      by = c("ensembl_gene_id" = "ensembl_gene_id"))

resDO_tb_significant <- resDO_tb_significant %>%
          left_join(tpm_counts, by = c("gene" = "ensembl_gene_id")) %>% 
          comb_de_result_table()

write_xlsx(list(T16.DE_date_of.day8 = resDO_tb_significant),
           "tables/T16.DE_date_of.day8.xlsx")

# Separate into up and down-regulated gene sets
sigDO_up <- rownames(resDO)[which(resDO$padj < 0.01 & resDO$log2FoldChange > 0)]
sigDO_down <- rownames(resDO)[which(resDO$padj < 0.01 & resDO$log2FoldChange < 0)]
```

\newpage

# Visualization

*Gene example*

```{r}
d <- plotCounts(dds, 
                gene = "ENSG00000130234", 
                intgroup = "response", 
                returnData = TRUE)

ggplot(d, aes(x = response, y = count)) + 
     geom_point(position = position_jitter(w = 0.1, h = 0)) +
     geom_text_repel(aes(label = rownames(d))) + 
     theme_bw(base_size = 10) +
     ggtitle("ACE2") +
     theme(plot.title = element_text(hjust = 0.5)) +
     scale_y_log10()
     
```

\newpage

```{r, fig.width = 10, fig.height = 10}
# Add a column for significant genes
resResponse_tb_vis <- resResponse_tb %>% mutate(threshold = padj < 0.01)

# don't show labels for lfc < 1.5
resResponse_tb_vis$symbol <- ifelse((abs(resResponse_tb_vis$log2FoldChange) > 1.5),   
                                resResponse_tb_vis$symbol, "") 


# don't show labels for non-significant genes
resResponse_tb_vis$symbol <- ifelse(resResponse_tb_vis$threshold,   
                                    resResponse_tb_vis$symbol, "") 

#fig4b
# suggested by Jennifer
dont_show <- c("C20orf204", "B4GALNT4", "SMIM34B", "ARHGAP33", 
               "ATRNL1", "CROCC2", "CEL", "FLG", "KIAA1549L", 
               "AHRR", "SLC25A27", "CPLX1", "SNX10", "STX11",
               "FUCA1", "ACP5", "MS4A4A", "GM2A", "TYROBP", "GBP5", 
               "SASH3", "CTSL", "TFEC", "DNASE2", "IGSF6", "IFI30", "LGALS3",
               "FCER1G", "VNN2", "SELPLG", "GPX1", "TAGAP", "PRDX1", "SLAMF8",
               "SSR4", "LST1", "VMO1", "OMD", "PRB2", "TBX21", "LIPA", "COMMD3-BMI1",
               "SOWAHD", "LGALS2", "SLAMF6", "SH2D1A", "ALDH2", "PTPRCAP")

resResponse_tb_vis[resResponse_tb_vis$symbol %in% dont_show,]$symbol <- ""

# no highlighting
#resResponse_tb_vis$highlight <- ifelse(resResponse_tb_vis$symbol %in% genes_to_show, 
#                                    "yes", "no")

# a handful of extreme genes are not shown, like PRSS33
fig4b <- ggplot(resResponse_tb_vis,
       aes(log2FoldChange, -log10(padj), label = symbol)) +
  geom_point(aes(colour = threshold)) +
  ggtitle("Response pCR vs non-pCR") +
  xlab("log2 fold change") + 
  ylab("-log10 adjusted p-value") +
  scale_x_continuous(limits = c(-10, 10)) +
  scale_y_continuous(limits = c(0, 7)) +
  theme(legend.position = "none",
        plot.title = element_text(size = rel(1.5), hjust = 0.5),
        axis.title = element_text(size = rel(1.25)),
        panel.grid.major = element_blank(), 
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        panel.background = element_blank()) +
  geom_label_repel(aes(label = symbol), 
                  data = subset(resResponse_tb_vis, log2FoldChange > 0),
                      size = 5, 
                      min.segment.length = 0,
                      segment.size = 0.1,
                      segment.color = "grey50",
                      direction     = "both",
                      max.overlaps = Inf,
                      segment.linetype = 5,
                      max.iter =  Inf,
                      hjust = "right",
                      xlim = c(3, 10),
                      force = 5,
                      force_pull = 0
                  ) +
  geom_label_repel(aes(label = symbol), 
                  data = subset(resResponse_tb_vis, log2FoldChange < 0),
                      size = 5, 
                      min.segment.length = 0,
                      segment.size = 0.1,
                      segment.color = "grey50",
                      direction     = "both",
                      max.overlaps = Inf,
                      segment.linetype = 5,
                      max.iter =  Inf,
                      hjust = "left",
                      xlim = c(-9, -4),
                      force = 5,
                      force_pull = 0
                  )

 # nudge_x = 9 - subset(resResponse_tb_vis, log2FoldChange > 0)$log2FoldChange,
#force_pull = 0, # attraction between label and point
 #                     force = 0.5 # repulsion between labels
#point.padding = 1,,
#                  point.size = NA,
#                  arrow = arrow(length = unit(0.015, "npc")),
#                  segment.linetype = 1,
#                  box.padding = 0.5,
#                  force_pull = 0,
#                  force = 1

saveRDS(fig4b, "data/fig4b.RDS")  
fig4b
```

\newpage

```{r}
# Add a column for significant genes
resER_tb <- resER_tb %>% mutate(threshold = padj < 0.01)

ggplot(resER_tb) +
  geom_point(aes(x = log2FoldChange, y = -log10(padj), colour = threshold)) +
  ggtitle("ER: Positive vs Negative") +
  xlab("log2 fold change") + 
  ylab("-log10 adjusted p-value") +
  scale_x_continuous(limits = c(-10,10)) +
  theme(legend.position = "none",
        plot.title = element_text(size = rel(1.5), hjust = 0.5),
        axis.title = element_text(size = rel(1.25)))
```

\newpage

```{r}
# Add a column for significant genes
resTP_tb <- resTP_tb %>% mutate(threshold = padj < 0.01)

ggplot(resTP_tb) +
  geom_point(aes(x = log2FoldChange, y = -log10(padj), colour = threshold)) +
  ggtitle("Tumor_percentage_high: High vs Low") +
  xlab("log2 fold change") + 
  ylab("-log10 adjusted p-value") +
  scale_x_continuous(limits = c(-10,10)) +
  theme(legend.position = "none",
        plot.title = element_text(size = rel(1.5), hjust = 0.5),
        axis.title = element_text(size = rel(1.25)))
```

\newpage

```{r}
# Add a column for significant genes
resDO_tb <- resDO_tb %>% mutate(threshold = padj < 0.01)

ggplot(resDO_tb) +
  geom_point(aes(x = log2FoldChange, y = -log10(padj), colour = threshold)) +
  ggtitle("Dafe of: 20180323 vs 20180228") +
  xlab("log2 fold change") + 
  ylab("-log10 adjusted p-value") +
  scale_x_continuous(limits = c(-10,10)) +
  theme(legend.position = "none",
        plot.title = element_text(size = rel(1.5), hjust = 0.5),
        axis.title = element_text(size = rel(1.25)))
```

\newpage

# Heatmaps
```{r, fig.width = 20, fig.height = 30}
# Create a matrix of normalized expression
sig_up <- resResponse_tb_significant %>% arrange(-log2FoldChange) %>% head(50) %>% pull(gene)
sig_down <- resResponse_tb_significant %>% arrange(log2FoldChange) %>% head(50) %>% pull(gene)
sig <- c(sig_up, sig_down)

row_annotation <- gene_symbol %>% 
                    as_tibble() %>% 
                    dplyr::filter(gene_id %in% sig)

plotmat <- txi$abundance[c(sig_up, sig_down),] %>% as.data.frame() %>% 
          rownames_to_column(var = "ensembl_gene_id") %>% 
          left_join(gene_symbol, by = c("ensembl_gene_id" = "gene_id")) %>% 
          drop_na(symbol)

plotmat$ensembl_gene_id <- NULL

plotmat <- plotmat %>% column_to_rownames(var = "symbol") %>% as.matrix()

# Color palette
heat.colors <- brewer.pal(6, "YlOrRd")

# Plot heatmap
pheatmap(plotmat, 
         scale = "row", 
         show_rownames = TRUE,
         border = FALSE,
         annotation = meta[, c("response"), drop = FALSE],
         main = "Top 50 Up- and Down- regulated genes in Response: pCR vs non-pCR",
         fontsize = 20)

```

\newpage

```{r, fig.width = 20, fig.height = 30}
# Create a matrix of normalized expression
sig_up <- resER_tb_significant %>% arrange(-log2FoldChange) %>% head(50) %>% pull(gene)
sig_down <- resER_tb_significant %>% arrange(log2FoldChange) %>% head(50) %>% pull(gene)
sig <- c(sig_up, sig_down)

row_annotation <- gene_symbol %>% 
                    as_tibble() %>% 
                    dplyr::filter(gene_id %in% sig)

plotmat <- txi$abundance[c(sig_up, sig_down),] %>% as.data.frame() %>% 
          rownames_to_column(var = "ensembl_gene_id") %>% 
          left_join(gene_symbol, by = c("ensembl_gene_id" = "gene_id")) %>% 
          drop_na(symbol)

plotmat$ensembl_gene_id <- NULL

plotmat <- plotmat %>% column_to_rownames(var = "symbol") %>% as.matrix()

# Color palette
heat.colors <- brewer.pal(6, "YlOrRd")

# Plot heatmap
pheatmap(plotmat, 
         scale = "row", 
         show_rownames = TRUE,
         border = FALSE,
         annotation = meta[, c("er"), drop = FALSE],
         main = "Top 50 Up- and Down- regulated genes in ER: positive vs negative",
         fontsize = 20)

```

\newpage

```{r, fig.width = 20, fig.height = 30}
# Create a matrix of normalized expression
sig_up <- resTP_tb_significant %>% arrange(-log2FoldChange) %>% head(50) %>% pull(gene)
sig_down <- resTP_tb_significant %>% arrange(log2FoldChange) %>% head(50) %>% pull(gene)
sig <- c(sig_up, sig_down)

row_annotation <- gene_symbol %>% 
                    as_tibble() %>% 
                    dplyr::filter(gene_id %in% sig)

plotmat <- txi$abundance[c(sig_up, sig_down),] %>% as.data.frame() %>% 
          rownames_to_column(var = "ensembl_gene_id") %>% 
          left_join(gene_symbol, by = c("ensembl_gene_id" = "gene_id")) %>% 
          drop_na(symbol)

plotmat$ensembl_gene_id <- NULL

plotmat <- plotmat %>% column_to_rownames(var = "symbol") %>% as.matrix()

# Color palette
heat.colors <- brewer.pal(6, "YlOrRd")

# Plot heatmap
pheatmap(plotmat, 
         scale = "row", 
         show_rownames = TRUE,
         border = FALSE,
         annotation = meta[, c("tumor_percentage_high"), drop = FALSE],
         main = "Top Up/Down-regulated genes in Tumor_percentage_high: high vs low",
         fontsize = 20)

```

\newpage

```{r, fig.width = 20, fig.height = 30}
# Create a matrix of normalized expression
sig_up <- resDO_tb_significant %>% arrange(-log2FoldChange) %>% head(50) %>% pull(gene)
sig_down <- resDO_tb_significant %>% arrange(log2FoldChange) %>% head(50) %>% pull(gene)
sig <- c(sig_up, sig_down)

row_annotation <- gene_symbol %>% 
                    as_tibble() %>% 
                    dplyr::filter(gene_id %in% sig)

plotmat <- txi$abundance[c(sig_up, sig_down),] %>% as.data.frame() %>% 
          rownames_to_column(var = "ensembl_gene_id") %>% 
          left_join(gene_symbol, by = c("ensembl_gene_id" = "gene_id")) %>% 
          drop_na(symbol)

plotmat$ensembl_gene_id <- NULL

plotmat <- plotmat %>% column_to_rownames(var = "symbol") %>% as.matrix()

# Color palette
heat.colors <- brewer.pal(6, "YlOrRd")

# Plot heatmap
pheatmap(plotmat, 
         scale = "row", 
         show_rownames = TRUE,
         border = FALSE,
         annotation = meta[, c("response"), drop = FALSE],
         main = "Top 50 Up- and Down- regulated genes in date_of: 20180323 vs 20180228",
         fontsize = 20)

```


```{r, include = FALSE}
#  prepares an expression profile for GSEA
#  http://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#Expression_Data_Formats
#  for GSEA it is important to report all genes - genome wide
#  hopefully cpms are better than logcpms
counts <- counts[rowSums(counts)>0,]
result_file <-  paste0("tables/8day.4gsea.txt")

counts_gsea <- counts %>% as.data.frame() %>% 
    rownames_to_column(var = "ensembl_gene_id") %>% 
    left_join(gene_symbol, by = c("ensembl_gene_id" = "gene_id")) %>% 
    dplyr::relocate(symbol)
#%>% 
#    dplyr::relocate(ensembl_gene_id)

colnames(counts_gsea)[1:2] <- c("NAME", "DESCRIPTION")

d <-duplicated(counts_gsea$NAME)
o <-  order(rowSums(counts_gsea[,rownames(meta)]),decreasing = T)
counts_gsea <- counts_gsea[o, ]
counts_gsea <- counts_gsea[!d, ]


samples_yes <- meta %>% dplyr::filter(response == "pCR") %>% row.names()
samples_no <- meta %>% dplyr::filter(response == "non-pCR") %>% row.names()

counts_gsea <- counts_gsea[,c("NAME", "DESCRIPTION", samples_yes, samples_no)]
# gsea now supports ENSEMBL_IDs
write_tsv(counts_gsea, result_file)
```

\newpage

# Functional analysis
## Biological Process (BP)
```{r, fig.width = 12, fig.height = 12}
bg_genes <- rownames(resResponse)

## Run GO enrichment analysis 
compgo_file <- "data/day8.compgo.up.RDS"
if (file.exists(compgo_file)){
    compGO <- readRDS(compgo_file)
}else{
    compGO <- enrichGO(gene = sigResponse_up,
                   universe = bg_genes,
                   keyType = "ENSEMBL",
                   OrgDb = "org.Hs.eg.db", 
                   ont = "BP", 
                   qvalueCutoff  = 0.05, 
                   pAdjustMethod = "BH",
                   readable = TRUE)
    saveRDS(compGO, compgo_file)
}

dotplot(compGO,
        showCategory = 20, 
        title = "GO (Biological Process) Enrichment \n Analysis for UP in pCR)",
        label_format = 20,
        font.size = 10)
```


```{r}
## Output results from GO analysis to a table

print("UP")
results_up <- data.frame(compGO@result) %>% dplyr::filter(p.adjust < 0.05)
nrow(results_up)
write_csv(results_up, "tables/T20.day8.GO_BP_UP.csv")
```

\newpage

```{r}
compgo_file <- "data/day8.compgo.down.RDS"
if (file.exists(compgo_file)){
    compGO <- readRDS(compgo_file)
}else{
    compGO <- enrichGO(gene = sigResponse_down,
                   universe = bg_genes,
                   keyType = "ENSEMBL",
                   OrgDb = "org.Hs.eg.db", 
                   ont = "BP", 
                   qvalueCutoff  = 0.05, 
                   pAdjustMethod = "BH",
                   readable = TRUE)
    saveRDS(compGO, compgo_file)
}


results_down <- data.frame(compGO@result) %>% dplyr::filter(p.adjust < 0.05)
print("Down")
nrow(results_down)
write_csv(results_down, "tables/T21.day8.GO_BP_DOWN.csv")
dotplot(compGO,
        showCategory = 20, 
        title = "GO (Biological Process) Enrichment \n Analysis for Down in pCR)",
        label_format = 20,
        font.size = 10)
```
\newpage

## Molecular Function (MF)
```{r, fig.width = 12, fig.height = 12}
bg_genes <- rownames(resResponse)

## Run GO enrichment analysis 
compgo_file <- "data/day8.compgo.up.mf.RDS"
if (file.exists(compgo_file)){
    compGO <- readRDS(compgo_file)
}else{
    compGO <- enrichGO(gene = sigResponse_up,
                   universe = bg_genes,
                   keyType = "ENSEMBL",
                   OrgDb = "org.Hs.eg.db", 
                   ont = "MF", 
                   qvalueCutoff  = 0.05, 
                   pAdjustMethod = "BH",
                   readable = TRUE)
    saveRDS(compGO, compgo_file)
}
# image pdf 12 x 12

## Output results from GO analysis to a table
print("UP")
results_up <- data.frame(compGO@result) %>% dplyr::filter(p.adjust < 0.05)
nrow(results_up)

write_csv(results_up, "tables/T22.day8.GO_MF_UP.csv")

dotplot(compGO,
        showCategory = 20, 
        title = "GO (Molecular Function) Enrichment \n Analysis for UP in pCR)",
        label_format = 20,
        font.size = 10)

```


\newpage

```{r, fig.width = 12, fig.height = 12}
compgo_file <- "data/day8.compgo.down.mf.RDS"
if (file.exists(compgo_file)){
    compGO <- readRDS(compgo_file)
}else{
  compGO <- enrichGO(gene = sigResponse_down,
                   universe = bg_genes,
                   keyType = "ENSEMBL",
                   OrgDb = "org.Hs.eg.db", 
                   ont = "MF", 
                   qvalueCutoff  = 0.05, 
                   pAdjustMethod = "BH",
                   readable = TRUE)
   saveRDS(compGO, compgo_file)
}

results_down <- data.frame(compGO@result) %>% dplyr::filter(p.adjust < 0.05)
print("Down")
nrow(results_down)
write_csv(results_up, "tables/T22.day8.GO_MF_DOWN.csv")

if (nrow(results_down) > 0){
  dotplot(compGO,
        showCategory = 20, 
        title = "GO (Molecular Function) Enrichment \n Analysis for DOWN in pCR)",
        label_format = 20,
        font.size = 10)
}

```
\newpage

## Cellular Compartment (CC)
```{r, fig.width = 12, fig.height = 12}
bg_genes <- rownames(resResponse)

## Run GO enrichment analysis 
compgo_file <- "data/day8.compgo.up.cc.RDS"
if (file.exists(compgo_file)){
    compGO <- readRDS(compgo_file)
}else{
  compGO <- enrichGO(gene = sigResponse_up,
                   universe = bg_genes,
                   keyType = "ENSEMBL",
                   OrgDb = "org.Hs.eg.db", 
                   ont = "CC", 
                   qvalueCutoff  = 0.05, 
                   pAdjustMethod = "BH",
                   readable = TRUE)
  saveRDS(compGO, compgo_file)
}

# image pdf 12 x 12

## Output results from GO analysis to a table
print("UP")
results_up <- data.frame(compGO@result) %>% dplyr::filter(p.adjust < 0.05)
nrow(results_up)

write_csv(results_up, "tables/T23.day8.GO_CC_UP.csv")


dotplot(compGO,
        showCategory = 20, 
        title = "GO (Cellular compartment) Enrichment \n Analysis for UP in pCR)",
        label_format = 20,
        font.size = 10)
```

\newpage

```{r, fig.width = 12, fig.height = 12}
compgo_file <- "data/day8.compgo.down.cc.RDS"
if (file.exists(compgo_file)){
    compGO <- readRDS(compgo_file)
}else{
    compGO <- enrichGO(gene = sigResponse_down,
                   universe = bg_genes,
                   keyType = "ENSEMBL",
                   OrgDb = "org.Hs.eg.db", 
                   ont = "CC", 
                   qvalueCutoff  = 0.05, 
                   pAdjustMethod = "BH",
                   readable = TRUE)
    saveRDS(compGO, compgo_file)
}

results_down <- data.frame(compGO@result) %>% dplyr::filter(p.adjust < 0.05)
print("Down")
nrow(results_down)
write_csv(results_up, "tables/T24.day8.GO_CC_DOWN.csv")

if (nrow(results_down) > 0){
dotplot(compGO,
        showCategory = 20, 
        title = "GO (Cellular compartment) Enrichment \n Analysis for DOWN in pCR)",
        label_format = 20,
        font.size = 10)
}
```

# R session
```{r}
sessionInfo()
```
